+++ internal/impl/sql/integration_test.go
@@ -528,7 +528,7 @@ processors:
 	})
 }
 
-func testSuite(t *testing.T, driver, dsn string, createTableFn func(string) error) {
+func testSuite(t *testing.T, driver, dsn string, createTableFn func(string) (string, error)) {
 	for _, fn := range []testFn{
 		testBatchProcessorBasic,
 		testBatchProcessorParallel,
@@ -540,7 +540,9 @@ func testSuite(t *testing.T, driver, dsn string, createTableFn func(string) erro
 		tableName, err := gonanoid.Generate("abcdefghijklmnopqrstuvwxyz", 40)
 		require.NoError(t, err)
 
-		require.NoError(t, createTableFn(tableName), tableName)
+		tableName, err = createTableFn(tableName)
+		require.NoError(t, err)
+
 		fn(t, driver, dsn, tableName)
 	}
 }
@@ -571,13 +573,13 @@ func TestIntegrationClickhouse(t *testing.T) {
 		}
 	})
 
-	createTable := func(name string) error {
+	createTable := func(name string) (string, error) {
 		_, err := db.Exec(fmt.Sprintf(`create table %s (
   "foo" String,
   "bar" Int64,
   "baz" String
 		) engine=Memory;`, name))
-		return err
+		return name, err
 	}
 
 	dsn := fmt.Sprintf("clickhouse://localhost:%s/", resource.GetPort("9000/tcp"))
@@ -591,7 +593,7 @@ func TestIntegrationClickhouse(t *testing.T) {
 			db = nil
 			return err
 		}
-		if err := createTable("footable"); err != nil {
+		if _, err := createTable("footable"); err != nil {
 			return err
 		}
 		return nil
@@ -626,13 +628,13 @@ func TestIntegrationOldClickhouse(t *testing.T) {
 		}
 	})
 
-	createTable := func(name string) error {
+	createTable := func(name string) (string, error) {
 		_, err := db.Exec(fmt.Sprintf(`create table %s (
   "foo" String,
   "bar" Int64,
   "baz" String
 		) engine=Memory;`, name))
-		return err
+		return name, err
 	}
 
 	dsn := fmt.Sprintf("tcp://localhost:%s/", resource.GetPort("9000/tcp"))
@@ -646,7 +648,7 @@ func TestIntegrationOldClickhouse(t *testing.T) {
 			db = nil
 			return err
 		}
-		if err := createTable("footable"); err != nil {
+		if _, err := createTable("footable"); err != nil {
 			return err
 		}
 		return nil
@@ -686,14 +688,14 @@ func TestIntegrationPostgres(t *testing.T) {
 		}
 	})
 
-	createTable := func(name string) error {
+	createTable := func(name string) (string, error) {
 		_, err := db.Exec(fmt.Sprintf(`create table %s (
   "foo" varchar(50) not null,
   "bar" integer not null,
   "baz" varchar(50) not null,
   primary key ("foo")
 		)`, name))
-		return err
+		return name, err
 	}
 
 	dsn := fmt.Sprintf("postgres://testuser:testpass@localhost:%s/testdb?sslmode=disable", resource.GetPort("5432/tcp"))
@@ -707,7 +709,7 @@ func TestIntegrationPostgres(t *testing.T) {
 			db = nil
 			return err
 		}
-		if err := createTable("footable"); err != nil {
+		if _, err := createTable("footable"); err != nil {
 			return err
 		}
 		return nil
@@ -751,14 +753,14 @@ func TestIntegrationMySQL(t *testing.T) {
 		}
 	})
 
-	createTable := func(name string) error {
+	createTable := func(name string) (string, error) {
 		_, err := db.Exec(fmt.Sprintf(`create table %s (
   "foo" varchar(50) not null,
   "bar" integer not null,
   "baz" varchar(50) not null,
   primary key ("foo")
 		)`, name))
-		return err
+		return name, err
 	}
 
 	dsn := fmt.Sprintf("testuser:testpass@tcp(localhost:%s)/testdb", resource.GetPort("3306/tcp"))
@@ -771,7 +773,7 @@ func TestIntegrationMySQL(t *testing.T) {
 			db = nil
 			return err
 		}
-		if err := createTable("footable"); err != nil {
+		if _, err := createTable("footable"); err != nil {
 			return err
 		}
 		return nil
@@ -811,14 +813,14 @@ func TestIntegrationMSSQL(t *testing.T) {
 		}
 	})
 
-	createTable := func(name string) error {
+	createTable := func(name string) (string, error) {
 		_, err := db.Exec(fmt.Sprintf(`create table %s (
   "foo" varchar(50) not null,
   "bar" integer not null,
   "baz" varchar(50) not null,
   primary key ("foo")
 		)`, name))
-		return err
+		return name, err
 	}
 
 	dsn := fmt.Sprintf("sqlserver://sa:"+testPassword+"@localhost:%s?database=master", resource.GetPort("1433/tcp"))
@@ -832,7 +834,7 @@ func TestIntegrationMSSQL(t *testing.T) {
 			db = nil
 			return err
 		}
-		if err := createTable("footable"); err != nil {
+		if _, err := createTable("footable"); err != nil {
 			return err
 		}
 		return nil
@@ -853,14 +855,14 @@ func TestIntegrationSQLite(t *testing.T) {
 		}
 	})
 
-	createTable := func(name string) error {
+	createTable := func(name string) (string, error) {
 		_, err := db.Exec(fmt.Sprintf(`create table %s (
   "foo" varchar(50) not null,
   "bar" integer not null,
   "baz" varchar(50) not null,
   primary key ("foo")
 		)`, name))
-		return err
+		return name, err
 	}
 
 	dsn := "file::memory:?cache=shared"
@@ -875,7 +877,7 @@ func TestIntegrationSQLite(t *testing.T) {
 			db = nil
 			return err
 		}
-		if err := createTable("footable"); err != nil {
+		if _, err := createTable("footable"); err != nil {
 			return err
 		}
 		return nil
@@ -914,14 +916,14 @@ func TestIntegrationOracle(t *testing.T) {
 		}
 	})
 
-	createTable := func(name string) error {
+	createTable := func(name string) (string, error) {
 		_, err := db.Exec(fmt.Sprintf(`create table %s (
   "foo" varchar(50) not null,
   "bar" integer not null,
   "baz" varchar(50) not null,
   primary key ("foo")
 		)`, name))
-		return err
+		return name, err
 	}
 
 	dsn := fmt.Sprintf("oracle://system:testpass@localhost:%s/XE", resource.GetPort("1521/tcp"))
@@ -937,7 +939,7 @@ func TestIntegrationOracle(t *testing.T) {
 			return err
 		}
 
-		if err := createTable("footable"); err != nil {
+		if _, err := createTable("footable"); err != nil {
 			return err
 		}
 		return nil
@@ -976,12 +978,18 @@ func TestIntegrationTrino(t *testing.T) {
 		}
 	})
 
-	selectTable := func(name string) error {
-		_, err := db.Exec(fmt.Sprintf(`SELECT * FROM %s;`, name))
-		return err
+	createTable := func(name string) (string, error) {
+		name = "memory.default." + name
+		_, err := db.Exec(fmt.Sprintf(`
+create table %s (
+  "foo" varchar,
+  "bar" integer,
+  "baz" varchar
+)`, name))
+		return name, err
 	}
 
-	dsn := fmt.Sprintf("https://trinouser:"+testPassword+"@localhost:%s", resource.GetPort("8080/tcp"))
+	dsn := fmt.Sprintf("http://trinouser:"+testPassword+"@localhost:%s", resource.GetPort("8080/tcp"))
 	require.NoError(t, pool.Retry(func() error {
 		db, err = sql.Open("trino", dsn)
 		if err != nil {
@@ -992,11 +1000,11 @@ func TestIntegrationTrino(t *testing.T) {
 			db = nil
 			return err
 		}
-		if err := selectTable("tpcds.tiny.customer"); err != nil {
+		if _, err := createTable("test"); err != nil {
 			return err
 		}
 		return nil
 	}))
 
-	testSuite(t, "trino", dsn, selectTable)
+	testSuite(t, "trino", dsn, createTable)
 }